added air behind the stator yoke and rotor backiron
init_geometry_2 adds a layer of air behind the stator yoke and rotor backiron.  Befor there was an asymptotic boundary condition at the steel boundary.  This assumption is not accurite when the steel is highly saturated.  Adding a layer of air and then the asymptotic boundary conidtion seems to fix that.

Also added a flag for hiding graphics.  makes it way faster.